[feat]カメラの定期作動
