+++ app/src/main/AndroidManifest.xml
@@ -23,6 +23,8 @@
             </intent-filter>
         </activity>
 
+        <activity android:name=".CameraActivity" />
+
     </application >
 
 </manifest>
